rename definitions.py to constants.py
The constants.py as a name makes more sense.
