+++ wakepy/core/__init__.py
@@ -5,13 +5,12 @@ See the public Python API at: https://wakepy.readthedocs.io/
 """
 
 from .activationresult import ActivationResult as ActivationResult
-from .activationresult import StageName as StageName
 from .calls import Call as Call
 from .calls import DbusMethodCall as DbusMethodCall
 from .configuration import CURRENT_SYSTEM as CURRENT_SYSTEM
-from .strenum import StrEnum as StrEnum
+from .constants import SystemName as SystemName
 from .dbus import BusType as BusType
 from .dbus import DbusAddress as DbusAddress
 from .dbus import DbusMethod as DbusMethod
-from .definitions import SystemName as SystemName
 from .method import Method as Method
+from .strenum import StrEnum as StrEnum
